IntegrationTest: load_test_thingpedia to typescript
